Dataset_1
This Excel file is part of a mini learning project designed to showcase and practice essential Excel formulas and functions. It includes a variety of commonly used formulas such as VLOOKUP, HLOOKUP, SUMIF, COUNTIF, and others. The project aims to demonstrate practical applications of these functions in data analysis and manipulation.

Feel free to explore and use this file for learning or reference purposes. Contributions or suggestions for improvement are always welcome!